¡Estilo WhatsApp implementado! 🚀
He refactorizado la interfaz de grabación:

Lógica Extraída: Creé un hook useVoiceRecorder para manejar la lógica de grabación de forma limpia.
Interfaz Dinámica:
Al pulsar el micrófono 🎙️, desaparece la barra de texto.
Aparece una barra blanca con:
🗑️ Papelera (izquierda): para cancelar.
⏱️ Temporizador + "Recording..." (centro).
➡️ Enviar (derecha): para detener y enviar.
Animaciones: Se añadieron transiciones suaves (animate-fade-in) para que el cambio de iconos se sienta natural.
