--- frontend/src/components/Buttons/ButtonMic.jsx
@@ -1,108 +1,0 @@
-import { useState, useRef, useEffect } from 'react';
-import PropTypes from 'prop-types';
-
-const ButtonMic = ({ onAudioSubmit }) => {
-    const [isRecording, setIsRecording] = useState(false);
-    const [recordingTime, setRecordingTime] = useState(0);
-    const mediaRecorderRef = useRef(null);
-    const audioChunksRef = useRef([]);
-    const timerRef = useRef(null);
-
-    useEffect(() => {
-        return () => {
-            if (timerRef.current) clearInterval(timerRef.current);
-        };
-    }, []);
-
-    const startRecording = async () => {
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-            const mediaRecorder = new MediaRecorder(stream);
-            mediaRecorderRef.current = mediaRecorder;
-            audioChunksRef.current = [];
-
-            mediaRecorder.ondataavailable = (event) => {
-                audioChunksRef.current.push(event.data);
-            };
-
-            mediaRecorder.onstop = () => {
-                const audioBlob = new Blob(audioChunksRef.current, { type: 'audio/webm' });
-                const reader = new FileReader();
-                reader.readAsDataURL(audioBlob);
-                reader.onloadend = () => {
-                    const base64Audio = reader.result;
-                    onAudioSubmit(base64Audio);
-                };
-                
-                // Stop all tracks
-                stream.getTracks().forEach(track => track.stop());
-            };
-
-            mediaRecorder.start();
-            setIsRecording(true);
-            
-            // Start Timer
-            setRecordingTime(0);
-            timerRef.current = setInterval(() => {
-                setRecordingTime(prev => {
-                    if (prev >= 60) { // Limit to 60 seconds
-                         stopRecording();
-                         return 60;
-                    }
-                    return prev + 1;
-                });
-            }, 1000);
-
-        } catch (error) {
-            console.error("Error accessing microphone:", error);
-            alert("Could not access microphone. Please allow permissions.");
-        }
-    };
-
-    const stopRecording = () => {
-        if (mediaRecorderRef.current && isRecording) {
-            mediaRecorderRef.current.stop();
-            setIsRecording(false);
-            if (timerRef.current) clearInterval(timerRef.current);
-        }
-    };
-
-    const formatTime = (seconds) => {
-        const mins = Math.floor(seconds / 60);
-        const secs = seconds % 60;
-        return `${mins}:${secs < 10 ? '0' : ''}${secs}`;
-    };
-
-    return (
-        <button
-            type="button"
-            onClick={isRecording ? stopRecording : startRecording}
-            className={`p-2 rounded-full transition-all duration-300 flex items-center gap-2 ${
-                isRecording 
-                ? "bg-red-500 hover:bg-red-600 animate-pulse text-white pr-4" 
-                : "text-gray-500 hover:text-red-500 hover:bg-gray-100"
-            }`}
-            title={isRecording ? "Stop Recording" : "Record Voice Message"}
-        >
-            {isRecording ? (
-                <>
-                     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="w-6 h-6">
-                        <path fillRule="evenodd" d="M4.5 7.5a3 3 0 0 1 3-3h9a3 3 0 0 1 3 3v9a3 3 0 0 1-3 3h-9a3 3 0 0 1-3-3v-9Z" clipRule="evenodd" />
-                    </svg>
-                    <span className="font-mono font-bold text-sm">{formatTime(recordingTime)}</span>
-                </>
-            ) : (
-                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="w-6 h-6">
-                    <path d="M8.25 4.5a3.75 3.75 0 1 1 7.5 0v8.25a3.75 3.75 0 1 1-7.5 0V4.5Z" />
-                    <path d="M6 10.5a.75.75 0 0 1 .75.75v1.5a5.25 5.25 0 1 0 10.5 0v-1.5a.75.75 0 0 1 1.5 0v1.5a6.751 6.751 0 0 1-6 6.709v2.291h3a.75.75 0 0 1 0 1.5h-7.5a.75.75 0 0 1 0-1.5h3v-2.291a6.751 6.751 0 0 1-6-6.709v-1.5A.75.75 0 0 1 6 10.5Z" />
-                </svg>
-            )}
-        </button>
-    );
-};
-
-ButtonMic.propTypes = {
-    onAudioSubmit: PropTypes.func.isRequired,
-};
-
-export default ButtonMic;
